Updated the Documentation of ianvs by replacing pcb-aoi with cloud-edge collaborative-inference-for-llm example
Updated the Documentation of ianvs by replacing pcb-aoi with cloud-edge collaborative-inference-for-llm example

Updates the Documentation of ianvs by replacing pcb-aoi with cloud-edge collaborative-inference-for-llm example

Updated the ianvs Quick Start guide. Replaced the PCB-AOI related content with cloud-edge-collaborative-inference-for-LLM.

Updated the ianvs Quick Start guide. Replaced the PCB-AOI related content with cloud-edge-collaborative-inference-for-LLM.

Signed-off-by: Aryan <nandaaryan823@gmail.com>

changes done on ianvs quick-start guide

Signed-off-by: Aryan <nandaaryan823@gmail.com>

how-to-use-ianvs-command-line updated to use cloud-edge-collaborative-inference-for-llm as an example instead of pcb-aoi

Signed-off-by: Aryan <nandaaryan823@gmail.com>

Cloud-Edge-Collaborative-Inference-For-LLM scenarion added to Scenarios section of docs with details of MMLU-5-Shot dataset

Signed-off-by: Aryan <nandaaryan823@gmail.com>

Joint Inference: Query-Routing Algorithm Added to algorithms section of ianvs documentation

Signed-off-by: Aryan <nandaaryan823@gmail.com>

benchmarking.yml file present in How to build simulation env updated to use cloud-edge inference for llm instead of pcb-aoi

Signed-off-by: Aryan <nandaaryan823@gmail.com>

testenv present in how-to-contribute-test-environments.md updated to use cloud-edge-collaborative-inference-for-llm as an example instead of pcb-aoi

Signed-off-by: Aryan <nandaaryan823@gmail.com>

how-to-contribute-algorithms updated to use cloud-edge-collaborative-inference-for-llm as an example as well

Signed-off-by: Aryan <nandaaryan823@gmail.com>

images folder removed. from docs/proposalsa/scenarios/cloud-edge-collaborative-inference

Signed-off-by: Aryan <nandaaryan823@gmail.com>

how-to-test-algorithms updated to include cloud-edge-collaborative-inference-for-llm example

Signed-off-by: Aryan <nandaaryan823@gmail.com>

leaderboard of cloud-edge-collaborative-inference-for-llm scenario added

Signed-off-by: Aryan <nandaaryan823@gmail.com>

Testing Joint Inference Learning in Cloud Edge Collaborative Inference for LLM Scenario with Ianvs-MMLU-5-shot dataset added

Signed-off-by: Aryan <nandaaryan823@gmail.com>

user_interfaces guides updated to use cloud-edge-collaborative-inference-for-llm as an example instead of pcb-aoi

Signed-off-by: Aryan <nandaaryan823@gmail.com>

index.rst updated to restructure leaderboards as per test-reports

Signed-off-by: Aryan <nandaaryan823@gmail.com>

cloud-edge-collaborative-inference-for-llm design image

Signed-off-by: Aryan <nandaaryan823@gmail.com>

cloud-edge-collaborative-inference-for-llm design image

Signed-off-by: Aryan <nandaaryan823@gmail.com>

cloud-robotics-lifelong-learning scenario added

Signed-off-by: Aryan <nandaaryan823@gmail.com>

dataset download instruction changed from hugging face to kaggle

dataset download instruction changed from hugging face to kaggle

Signed-off-by: Aryan <nandaaryan823@gmail.com>

version in how to install updated